SPDX ID and comments review.
