+++ rust_dev_preview/cross_service/rest_ses/src/configuration.rs
@@ -1,3 +1,8 @@
+/*
+ * Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+ * SPDX-License-Identifier: Apache-2.0.
+ */
+
 //! Environment and settings loading.
 //!
 //! Loads environment specific details for ARNs, tcp listeners, etc.
