+++ rust_dev_preview/cross_service/rest_ses/src/lib.rs
@@ -1,3 +1,8 @@
+/*
+ * Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+ * SPDX-License-Identifier: Apache-2.0.
+ */
+
 pub mod client;
 pub mod configuration;
 pub mod healthz;
