+++ rust_dev_preview/cross_service/rest_ses/src/main.rs
@@ -1,3 +1,8 @@
+/*
+ * Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+ * SPDX-License-Identifier: Apache-2.0.
+ */
+
 //! Main that loads environments & prepares clients, and hands them to `startup`.
 use std::net::TcpListener;
