+++ rust_dev_preview/cross_service/rest_ses/src/report.rs
@@ -1,3 +1,8 @@
+/*
+ * Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+ * SPDX-License-Identifier: Apache-2.0.
+ */
+
 //! The `/items:report` endpoint.
 use crate::{
     client::{Email, RdsClient, SesClient},
