+++ rust_dev_preview/cross_service/rest_ses/src/work_item/collection.rs
@@ -1,3 +1,8 @@
+/*
+ * Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+ * SPDX-License-Identifier: Apache-2.0.
+ */
+
 //! Provides scoped HTTP endpoints for a REST WorkItem collection.
 //! This includes the common REST HTTP endpoints, as well as RPC-like endpoints.
 //!
