+++ rust_dev_preview/cross_service/rest_ses/tests/api/work_item/happy.rs
@@ -1,3 +1,8 @@
+/*
+ * Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+ * SPDX-License-Identifier: Apache-2.0.
+ */
+
 //! End-to-end tests for the Rest SES happy-path API.
 //! This uses real aws environments and real aws services.
 //! It will result in billing for an AWS account!
