+++ rust_dev_preview/cross_service/rest_ses/tests/api/work_item/mod.rs
@@ -1,3 +1,8 @@
+/*
+ * Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+ * SPDX-License-Identifier: Apache-2.0.
+ */
+
 use fake::{
     faker::{lorem::en::Sentence, name::en::Name},
     Fake,
